+++ example_04_pattern_analysis/download_data.sh
@@ -5,17 +5,21 @@
 
 # usage: "sh ./prepare_data.sh"
 
-echo -e "\nPreparing data for sample fold analysis."
+if [ -f WT_BR1_comp_hOH7_KSHV_hybrids_ua.vienna ]; then
+    printf "Data already downloaded.\n"
+else
+    printf "\nPreparing data for sample fold analysis."
 
-echo -e "\nCopying data to analysis directory."
-cp -v ../ref_data/WT_BR1_comp_hOH7_KSHV_hybrids_ua.hyb.gz \
-      ../ref_data/WT_BR1_comp_hOH7_KSHV_hybrids_ua.vienna.gz \
-      ./
-  
+      printf "\nCopying data to analysis directory."
+      cp -v ../ref_data/WT_BR1_comp_hOH7_KSHV_hybrids_ua.hyb.gz \
+            ../ref_data/WT_BR1_comp_hOH7_KSHV_hybrids_ua.vienna.gz \
+            ./
 
-# Unzip (gunzip) individual hyb-format data-files.
-echo -e "\nUnpacking gz-compressed hyb data files...\n"
-gunzip WT_BR1_comp_hOH7_KSHV_hybrids_ua.hyb.gz \
-       WT_BR1_comp_hOH7_KSHV_hybrids_ua.vienna.gz
 
-echo -e "Done.\n"
+      # Unzip (gunzip) individual hyb-format data-files.
+      printf "\nUnpacking gz-compressed hyb data files...\n"
+      gunzip WT_BR1_comp_hOH7_KSHV_hybrids_ua.hyb.gz \
+            WT_BR1_comp_hOH7_KSHV_hybrids_ua.vienna.gz
+fi
+
+printf "Done.\n"
